requirements.txt
pipecat-ai[daily,openai,silero]